requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
gTTS==2.3.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
loguru==0.6.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.0
opencv-python==4.7.0.72
pkg_resources==0.0.0
PyAudio==0.2.13
Pygments==2.14.0
pyzmq==25.0.0
requests==2.28.2
rfc3986==1.5.0
rich==13.3.2
sniffio==1.3.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
